requirements.txt
asgiref==3.7.1
biopython==1.81
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2.1
django-crispy-forms==2.0
django-jquery==3.1.0
djangorestframework==3.14.0
docopt==0.6.2
fontawesomefree==6.4.0
idna==3.4
joblib==1.2.0
numpy==1.24.3
packaging==23.1
pandas==1.4.2
pipreqs==0.4.13
plotly==5.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.1.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
whitenoise==6.4.0
xgboost==1.4.2
yarg==0.1.9
